semantic_search/requirements.txt
scikit-learn
numpy
tqdm
gitpython
python-dotenv
